moved initial implementation of visual techniques to its own module
