+++ clayrs/content_analyzer/field_content_production_techniques/visual_techniques/__init__.py
@@ -1,0 +1,1 @@
+from .visual_content_techniques import SkImageHogDescriptor, SkImageCannyEdgeDetector
